+++ lib/modules/navigation/navigation_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: navigation
-// Updated at: 2026-08-13 12:26:13
-// Commit: test(navigation): add unit tests for navigation business logic
+// Updated at: 2026-08-13 16:35:24
+// Commit: feat(navigation): add bi-directional event bus handler for navigation
 
 class navigationComponentState {
-  final String commitRef = '3570';
-  final String lastMessage = 'test(navigation): add unit tests for navigation business logic';
-  final DateTime timestamp = DateTime.parse('2026-08-13T12:26:13+05:00');
+  final String commitRef = '3581';
+  final String lastMessage = 'feat(navigation): add bi-directional event bus handler for navigation';
+  final DateTime timestamp = DateTime.parse('2026-08-13T16:35:24+05:00');
 }